rename docs -> hashreleaseserver
